% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mustard.R
\name{bin_cells}
\alias{bin_cells}
\title{Cell binning}
\usage{
bin_cells(expr, pseudotime, cellanno, interval_len = 50)
}
\arguments{
\item{expr}{The normalized gene expression matrix. Rows represent genes and columns represent cells}

\item{pseudotime}{The vector of user-provided pseudotime values}

\item{cellanno}{The vector indicating which sample each cell belongs to}

\item{interval_len}{Number of consecutive intervals (50 by default)}
}
\value{
A list including "expr", "pseudotime", and "cellanno" of the metacells
}
\description{
Cell binning
}
