% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mustard.R
\name{est_test_sample}
\alias{est_test_sample}
\title{Estimate sample loadings from the testing samples}
\usage{
est_test_sample(expr_test, pseudotime_test, cellanno_test = NULL, res_tempted)
}
\arguments{
\item{expr_test}{Same format as "expr" but for testing samples}

\item{pseudotime_test}{Same format as "pseudotime" but for testing samples}

\item{cellanno_test}{Same format as "cellanno" but for testing samples}

\item{res_tempted}{The list from \code{\link{mustard}} run on the training samples}
}
\value{
Sample loading matrix from the testing samples
}
\description{
Estimate sample loadings from the testing samples
}
