% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mustard.R
\name{mustard}
\alias{mustard}
\title{Trajectory-guided dimension reduction for multi-sample single-cell RNA-seq data}
\usage{
mustard(
  expr,
  pseudotime,
  cellanno = NULL,
  interval = NULL,
  r = 3,
  resolution = 101,
  smooth = 0.001,
  maxiter = 20,
  epsilon = 1e-04
)
}
\arguments{
\item{expr}{The normalized and standardized gene expression matrix. Rows represent genes and columns represent cells}

\item{pseudotime}{The vector of user-provided pseudotime values}

\item{cellanno}{The vector indicating which sample each cell belongs to}

\item{interval}{Range of pseudotime (range of the original pseudotime by default)}

\item{r}{Number of components to decompose into (3 by default)}

\item{resolution}{Number of pseudotime values to evaluate in the temporal loading function (101 by default). It does not affect the sample or gene loading vector}

\item{smooth}{Smoothing parameter for RKHS norm (1e-3 by default). Larger value means smoother temporal loading functions}

\item{maxiter}{Maximum number of iterations (20 by default)}

\item{epsilon}{Convergence criteria for difference between iterations (1e-4 by default)}
}
\value{
A list including the estimated loadings and explained variances
\describe{
\item{A.hat}{Sample loading matrix}
\item{B.hat}{Gene loading matrix}
\item{Phi.hat}{Temporal loading matrix}
\item{time}{The pseudotime values where the temporal loading function is evaluated}
\item{Lambda}{Eigenvalue vector}
\item{r.square}{Variance explained by each component, which is the R-squared of the linear regression of the vectorized temporal tensor against the vectorized low-rank reconstruction using each component}
\item{accum.r.square}{Variance explained by the top components accumulated, which is the R-squared of the linear regression of the vectorized temporal tensor against the vectorized low-rank reconstruction using the top components}
}
}
\description{
Trajectory-guided dimension reduction for multi-sample single-cell RNA-seq data
}
