% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mustard.R
\name{plot_gene_loading}
\alias{plot_gene_loading}
\title{Visualize the gene loadings}
\usage{
plot_gene_loading(res, i = 1, j = 2, ...)
}
\arguments{
\item{res}{The list from \code{\link{mustard}}}

\item{i}{The component to plot in the x-axis}

\item{j}{The component to plot in the y-axis}

\item{...}{Arguments in \code{ggplot2::geom_point(...)}}
}
\value{
A ggplot2 plot
}
\description{
Visualize the gene loadings
}
