% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mustard.R
\name{plot_sample_loading}
\alias{plot_sample_loading}
\title{Visualize the sample loadings}
\usage{
plot_sample_loading(res, meta, group, i = 1, j = 2, ...)
}
\arguments{
\item{res}{The list from \code{\link{mustard}}}

\item{meta}{The sample-level metadata}

\item{group}{The phenotype of interest}

\item{i}{The component to plot in the x-axis}

\item{j}{The component to plot in the y-axis}

\item{...}{Arguments in \code{ggplot2::geom_point(...)}}
}
\value{
A ggplot2 plot
}
\description{
Visualize the sample loadings
}
