% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mustard.R
\name{plot_temporal_loading}
\alias{plot_temporal_loading}
\title{Visualize the temporal loadings}
\usage{
plot_temporal_loading(res, r = NULL, ...)
}
\arguments{
\item{res}{The list from \code{\link{mustard}}}

\item{r}{The vector of components to plot}

\item{...}{Arguments in \code{ggplot2::geom_line(...)}}
}
\value{
A ggplot2 plot
}
\description{
Visualize the temporal loadings
}
