% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mustard.R
\name{scale_genes}
\alias{scale_genes}
\title{Gene scaling}
\usage{
scale_genes(expr, scale.max = 10)
}
\arguments{
\item{expr}{The normalized gene expression matrix. Rows represent genes and columns represent cells}

\item{scale.max}{Maximum value to return for scaled expression (10 by default)}
}
\value{
Same format as "expr"
}
\description{
Gene scaling
}
